# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  codeberg_org_azzet_azzetbe_internal_shared.APIResponse:
    properties:
      data: {}
      meta: {}
      success:
        type: boolean
    type: object
  codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIError'
      success:
        type: boolean
    type: object
...
  codeberg_org_azzet_azzetbe_internal_workspace.CreateWorkspaceRequest:
    description: Create a workspace from an existing entity
    properties:
      entity_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  codeberg_org_azzet_azzetbe_internal_workspace.MessageResponse:
    description: Simple message response
    properties:
      message:
        example: Operation successful
        type: string
    type: object
...
  codeberg_org_azzet_azzetbe_internal_workspace.SetCounterpartyAliasRequest:
    properties:
      custom_alias:
        type: string
      entity_id:
        type: string
    type: object
...
  codeberg_org_azzet_azzetbe_internal_workspace.WorkspaceResponse:
    description: Workspace information
    properties:
      created_at:
        example: "2026-05-20T10:00:00Z"
        type: string
      entity_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      entity_name:
        example: PT Maju Jaya
        type: string
      entity_type:
        example: BADAN_USAHA
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      plan_name:
        example: Professional
        type: string
      role:
        example: PEMILIK
        type: string
      subscription_status:
        example: active
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Enterprise-grade accounting, tax, and finance platform. User routes
    use Bearer JWT from /auth/login. Workspace-scoped routes require X-Workspace-ID
    header. Admin routes under /admin use admin JWT from /admin/auth/login.
  title: Azzet API
  version: "1.0"
paths:
...
  /workspaces:
    get:
      description: Returns all workspaces the authenticated user has access to.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_workspace.WorkspaceResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my workspaces
      tags:
      - Workspaces
    post:
      consumes:
      - application/json
      description: Create a workspace from an existing entity. The user becomes PEMILIK
        (owner).
      parameters:
      - description: Workspace data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_workspace.CreateWorkspaceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_workspace.WorkspaceResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create workspace
      tags:
      - Workspaces
  /workspaces/counterparties:
    get:
      description: Returns all counterparties (PELANGGAN + VENDOR) of the current
        workspace.
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_workspace.CounterpartyResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List counterparties
      tags:
      - Workspace Counterparties
    post:
      consumes:
      - application/json
      description: Add a counterparty (customer/vendor) to workspace. Creates shadow
        entity if entity_id is not provided.
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Counterparty data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_workspace.AddCounterpartyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_workspace.CounterpartyResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add counterparty
      tags:
      - Workspace Counterparties
  /workspaces/counterparties/aliases:
    get:
      parameters:
      - description: Workspace entity ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
      security:
      - BearerAuth: []
      summary: List counterparty aliases
      tags:
      - Workspace
    post:
      consumes:
      - application/json
      parameters:
      - description: Workspace entity ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Alias request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_workspace.SetCounterpartyAliasRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
      security:
      - BearerAuth: []
      summary: Set custom alias for counterparty
      tags:
      - Workspace
  /workspaces/counterparties/aliases/{entity_id}:
    delete:
      parameters:
      - description: Workspace entity ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Counterparty entity ID
        in: path
        name: entity_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete counterparty alias
      tags:
      - Workspace
  /workspaces/counterparties/search:
    get:
      parameters:
      - description: Workspace entity ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
      security:
      - BearerAuth: []
      summary: Search counterparties by name
      tags:
      - Workspace
...
  /workspaces/members:
    get:
      description: Returns all members (PEMILIK + KARYAWAN) of the current workspace.
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_workspace.MemberResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List workspace members
      tags:
      - Workspace Members
  /workspaces/members/{id}:
    delete:
      description: Remove a member from the workspace. Cannot remove the owner.
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Relation ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_workspace.MessageResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove workspace member
      tags:
      - Workspace Members
    patch:
      consumes:
      - application/json
      description: Update a member's role, alias, or status.
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Relation ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_workspace.UpdateMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_workspace.MessageResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update workspace member
      tags:
      - Workspace Members
  /workspaces/roles:
    get:
      description: Returns all custom roles for the current workspace.
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_workspace.RoleResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List workspace roles
      tags:
      - Workspaces
    post:
      consumes:
      - application/json
      description: Create a new custom role with specific permissions for the workspace.
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Role data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_workspace.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_workspace.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a custom workspace role
      tags:
      - Workspace Roles
  /workspaces/roles/{id}:
    delete:
      description: Delete a custom role. System roles cannot be deleted.
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_workspace.MessageResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a workspace role
      tags:
      - Workspace Roles
    patch:
      consumes:
      - application/json
      description: Update name, description, or permissions of a custom role.
      parameters:
      - description: Workspace ID
        in: header
        name: X-Workspace-ID
        required: true
        type: string
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Role data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_workspace.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_workspace.MessageResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/codeberg_org_azzet_azzetbe_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a workspace role
      tags:
      - Workspace Roles
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"